+++ rasa_core/server.py
@@ -609,6 +609,15 @@ def create_app(agent,
             "tracker": tracker.current_state(verbosity)
         })
 
+    @app.route("/parse",
+               methods=['POST', 'OPTIONS'])
+    @requires_auth(app, auth_token)
+    @cross_origin(origins=cors_origins)
+    @ensure_loaded_agent(agent)
+    def parse():
+        request_params = request.get_json()
+        return jsonify(agent.interpreter.parse(request_params.get("q")))
+
     return app