+++ tests/test_tracker_stores.py
@@ -49,7 +49,7 @@ def test_tracker_store_endpoint_config_loading():
     cfg = utils.read_endpoint_config(DEFAULT_ENDPOINTS_FILE, "tracker_store")
 
     assert cfg == EndpointConfig.from_dict({
-        "store_type": "redis",
+        "type": "redis",
         "url": "localhost",
         "port": 6379,
         "db": 0,
@@ -96,7 +96,7 @@ def test_tracker_store_from_string(default_domain):
 def test_tracker_store_from_invalid_module(default_domain):
     endpoints_path = "data/test_endpoints/custom_tracker_endpoints.yml"
     store_config = utils.read_endpoint_config(endpoints_path, "tracker_store")
-    store_config.store_type = "a.module.which.cannot.be.found"
+    store_config.type = "a.module.which.cannot.be.found"
 
     tracker_store = TrackerStore.find_tracker_store(default_domain,
                                                     store_config)
@@ -107,7 +107,7 @@ def test_tracker_store_from_invalid_module(default_domain):
 def test_tracker_store_from_invalid_string(default_domain):
     endpoints_path = "data/test_endpoints/custom_tracker_endpoints.yml"
     store_config = utils.read_endpoint_config(endpoints_path, "tracker_store")
-    store_config.store_type = "any string"
+    store_config.type = "any string"
 
     tracker_store = TrackerStore.find_tracker_store(default_domain,
                                                     store_config)