+++ tests/test_utils.py
@@ -90,7 +90,7 @@ def test_endpoint_config():
                     "password": "pass"},
         token="mytoken",
         token_name="letoken",
-        store_type="redis",
+        store="redis",
         port=6379,
         db=0,
         password="password",